WIP/TLBGen: add tlbgen
Add enough bindings for the COM interfaces to load a specified type
library.  This will allow Swift to use hand written COM bindings to
process type libraries and serve as a MIDL compiler to generate the
bindings to COM interfaces.  The result of this should allow a some
what more idiomatic wrapping to be auto-generated rather than manually
written.