IDXGIOutput: correct `GetDisplayModeList`
When `IDXGIOutput.GetDisplayModeList` is invoked with a buffer, the
`NumNodes` parameter is expected to be a non-zero value which indicates
the number of entries that can be held.  We would overwrite this value
with 0, which resulted in an error being reported:
~~~
0x887a0003 - The caller did not supply a sufficiently large buffer.
~~~
Avoid shadowing the variable to avoid this.  Introduce a test suite to
add coverage for this case.

Fixes: #59