SwiftCOM: update interfaces for newer SDK
Many of the vtable issues have now been corrected.  Unfortunately, it is
unclear which SDK version fixed the issue and this makes providing a
backwards compatible implementation difficult.  I would be open to
patches to support the older SDKs while retaining the ability to use the
newer SDKs as well.  The newer SDKs are more critical for CI though as
the builders tend to have the newest SDKs installed.